Implemented the core functionality
You can now build the tree in a declarative fashion using DWidgets (analogous to StatelessWidgets in Flutter) and DNodes (wrappers for Godot Node types with a matching class hierarchy, analogous to other built-in Flutter types which extend Widget). The tree can be rebuilt and an attempt will be made to reuse existing nodes in the tree. A _Key can be specified so that the right node will be found. _RecreateNode can specified to omit this process and forcefully create a new node. Old nodes, unused by the rebuilt tree, will be freed.